Adding the refactored code
